+++ completions.sh
@@ -3,13 +3,40 @@ _tbot()
     local cur prev words cword
     _init_completion || return
 
-    if [[ "$prev" == -T ]]; then
+    # Collect testcase path arguments
+    local index=0
+    local path_args=()
+    local curdir="$(pwd)"
+    local workdir="${curdir}"
+    while [[ $index -lt ${#words[@]} ]]; do
+        local current_word="${words[$index]}"
+
+        if [[ "$current_word" == @(-T|-t) ]]; then
+            local index=$(($index + 1))
+            path_args=("${path_args[@]}" "$current_word" "${words[$index]}")
+        elif [[ "$current_word" == -C ]]; then
+            local index=$(($index + 1))
+            path_args=("${path_args[@]}" "$current_word" "${words[$index]}")
+            workdir="${words[$index]}"
+        fi
+
+        local index=$(($index + 1))
+    done
+
+    if [[ "$prev" == -C ]]; then
         _filedir -d
         return
     fi
 
+    if [[ "$prev" == -T ]]; then
+        cd "${workdir}" && _filedir -d
+        cd "${curdir}"
+        return
+    fi
+
     if [[ "$prev" == @(-t|-b|-l|--board|--lab) ]]; then
-        _filedir py
+        cd "${workdir}" && _filedir py
+        cd "${curdir}"
         # Remove __pycache__ for convenience
         COMPREPLY=("${COMPREPLY[@]/*__pycache__*/}")
         return
@@ -20,22 +47,8 @@ _tbot()
         return
     fi
 
-    # Collect testcase path arguments
-    local index=0
-    local path_args=()
-    while [[ $index -lt ${#words[@]} ]]; do
-        local current_word="${words[$index]}"
-
-        if [[ "$current_word" == @(-T|-t) ]]; then
-            local index=$(($index + 1))
-            path_args=("${path_args[@]}" "$current_word" "${words[$index]}")
-        fi
-
-        local index=$(($index + 1))
-    done
-
     if [[ "$cur" == -* ]]; then
-        COMPREPLY=( $( compgen -W '-h -b -l -T -t -f -v -q -s -i
+        COMPREPLY=( $( compgen -W '-h -b -l -T -t -f -v -q -s -i -C
             --help
             --board
             --lab