+++ tbot/machine/linux/ssh/machine.py
@@ -109,7 +109,9 @@ class SSHMachine(linux.LinuxMachine):
             # Careful, password WILL BE LEAKED!
             cmd = ["sshpass", "-p", authenticator.password, "ssh"]
         elif isinstance(authenticator, auth.PrivateKeyAuthenticator):
-            cmd = ["ssh", "-o", "BatchMode=yes", "-i", str(authenticator.key)]
+            cmd = ["ssh", "-o", "BatchMode=yes", "-i", str(authenticator.key_file)]
+        elif isinstance(authenticator, auth.NoneAuthenticator):
+            cmd = ["ssh", "-o", "BatchMode=yes"]
         else:
             raise RuntimeError(f"{authenticator!r} is not supported for SSH hosts!")