+++ tbot/tc/selftest/tc/__init__.py
@@ -5,6 +5,7 @@ from tbot.machine import linux
 
 from .git import *  # noqa: F403
 from .shell import *  # noqa: F403
+from .build import *  # noqa: F403
 
 
 @tbot.testcase
@@ -14,7 +15,9 @@ def selftest_tc(lab: typing.Optional[linux.LabHost] = None,) -> None:
         tc.testsuite(
             selftest_tc_git_checkout,  # noqa: F405
             selftest_tc_git_am,  # noqa: F405
+            selftest_tc_git_apply,  # noqa: F405
             selftest_tc_git_bisect,  # noqa: F405
             selftest_tc_shell_copy,  # noqa: F405
+            selftest_tc_build_toolchain,  # noqa: F405
             lab=lh,
         )