+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Fixed
 - timeout in `read_until_prompt` sometimes being negative
+- log missing bootlog for U-Boot if no autoboot is configured
 
 
 ## [0.6.3] - 2018-11-28