+++ CHANGELOG.md
@@ -3,6 +3,26 @@
 ## [Unreleased]
 
 
+## [0.6.4] - 2018-12-05
+### Added
+- `-C` parameter to allow setting a different working directory
+- `${TBOTPATH}` environment variable to provide additional testcase
+  directories; `TBOTPATH` is a `:` separated list of directorie
+- `auth.NoneAuthenticator`: Authenticator without key nor password, useful
+  if ssh can figure authentication out by itself (eg with ssh-agent)
+- `GitRepository.apply`: Apply patches without committing the changes
+
+### Changed
+- `linux.shell`: Shells now have a `shell.command`, which allows specifying
+  a command to run the shell.  For example, bash is run using `bash --norc`
+  now.
+
+### Fixed
+- timeout in `read_until_prompt` sometimes being negative
+- log missing bootlog for U-Boot if no autoboot is configured
+- tbot ignored ssh host aliases (ref #8)
+
+
 ## [0.6.3] - 2018-11-28
 ### Added
 - `tbot.log.warning`: Print a warning message